requirements.txt
streamlit
langchain-community
langchain-openai
sqlalchemy
python-dotenv
matplotlib
pandas
plotly
